CR - Move navigation constants closer to component
