+++ apps/filecoin-site/src/app/_components/Navigation/DesktopNavigation.tsx
@@ -1,10 +1,11 @@
 'use client'
 
-import { headerNavigation } from '@/constants/navigation'
 
 import { NavigationLanguageToggle } from './NavigationLanguageToggle'
 import { NavigationMainLink } from './NavigationMainLink'
 import { NavigationMenu } from './NavigationMenu'
+
+import { headerNavigation } from '@/_components/Navigation/constants/navigation'
 
 export function DesktopNavigation() {
   return (
