+++ apps/filecoin-site/src/app/_components/Navigation/MobileNavigation.tsx
@@ -8,7 +8,6 @@ import { clsx } from 'clsx'
 import { IconButton } from '@filecoin-foundation/ui/IconButton'
 import { SlideOver } from '@filecoin-foundation/ui/SlideOver'
 
-import { internalNavigationItems } from '@/constants/navigation'
 
 import { useBackgroundVariant } from '@/components/BackgroundVariantContext'
 import { backgroundVariants } from '@/components/Section'
@@ -16,6 +15,8 @@ import { backgroundVariants } from '@/components/Section'
 import { variantMapping } from './constants/variantMapping'
 import { HomeLogoIconLink } from './HomeLogoIconLink'
 import { NavigationMainLink } from './NavigationMainLink'
+
+import { internalNavigationItems } from '@/_components/Navigation/constants/navigation'
 
 export function MobileNavigation() {
   const [open, setOpen] = useState(false)
