+++ apps/filecoin-site/src/app/_components/Navigation/NavigationMenu.tsx
@@ -4,11 +4,12 @@ import { clsx } from 'clsx'
 
 import { NavigationMenu as SharedNavigationMenu } from '@filecoin-foundation/ui/NavigationMenu'
 
-import type { LabelledExpandedNavItems } from '@/constants/navigation'
 
 import { desktopStyle } from './NavigationMainLink'
 import { NavigationMenuPanel } from './NavigationMenuPanel'
 import { useIsNavigationMenuActive } from './useIsNavigationMenuActive'
+
+import type { LabelledExpandedNavItems } from '@/_components/Navigation/constants/navigation'
 
 export function NavigationMenu({ label, items }: LabelledExpandedNavItems) {
   const isActive = useIsNavigationMenuActive(items)
