+++ apps/filecoin-site/src/app/_components/Navigation/NavigationMenuPanel.tsx
@@ -2,13 +2,14 @@
 
 import { clsx } from 'clsx'
 
-import { type LabelledExpandedNavItems } from '@/constants/navigation'
 
 import { useBackgroundVariant } from '@/components/BackgroundVariantContext'
 import { backgroundVariants } from '@/components/Section'
 
 import { variantMapping } from './constants/variantMapping'
 import { NavigationMenuLink } from './NavigationMenuLink'
+
+import { type LabelledExpandedNavItems } from '@/_components/Navigation/constants/navigation'
 
 const roundedStyle = 'rounded-2xl'
